change icon from ico to png
we notice that only few browsers(which are rarelly used,or outdated,or
used by people who would never use Exsimple,or even fanaticals).
ico is a outdated file-format,it can be 10 times larger to a same size
png.
And,I change the icon image itself,too.
The Original one was my avatar on Screeps,which is generated by
Screep,and when think of it,it's not so good to use it here.So I remake
it with Java,and now the picture is my own artificial,and I hava the
right to use it now.